Fix message interface errors in log4j2 integration. (#367)
Apparently getFormat() can (and does) return an empty string in some
cases. This would cause a `null` format string with parameters and
display as an Event/JSON issue in the Sentry UI. This fixes the log4j2
integration to only include the parameterized message information if
getFormat() returns something useful.