Issue #708: add a resource loader interface to target Android slow loading of jar files
The Lookup class will try to load Sentry properties from a classpath resource file by default.
This method is VERY slow on Android (on some older devices specifically).

Add a Sentry.init overload which accepts a ResourceLoader class to let Android users
specify a delegate that retrieves the properties file from the assets, which is way
faster on Android.