Optimize DuplicateEventDetectionEventProcessor performance.
Using WeakHashMap to store objects opens up a possibility that the map will grow to thousands of entries and causes performance loss when searching for stored exception. This may happen either if outside of the SDK code references to exceptions are kept or if GC does not run for long time.

Changing the implementation to `ConcurrentLinkedDeque` gives the event processor control over how many exceptions are stored in the memory.