Simplify configuring Logback integration when DSN is not set.
When an environment property that is referenced in logback.xml is not set, Logback sets its value to PROPERTY_NAME_IS_UNDEFINED (not a `null` like it could have been expected). This complicates setting up SentryLogback appender configuration where the integration should be disabled in environments that do not have this property.

Before this change, users had to do:

<options>
  <dsn>${PROPERTY:- }
</options>

With this change:

<options>
  <dsn>${PROPERTY}
</options>