feat(release): Delegate version bumping to Craft (#1518)
At the moment, `buildVersionCode` and `versionName` have to be updated manually before doing a release. This PR adds a version bumping script, run by Craft on [`prepare`](https://github.com/getsentry/craft#pre-release-command).

- `buildVersionCode` will be incremented by 1.
- `versionName` will be the new version to prepare by Craft. For example, `craft prepare 6.0.0` will set `versionName=6.0.0`, and `craft prepare 6.1.0-alpha.1` will set `versionName=6.1.0-alpha.1`.

Note [Changelog Policies](6.1.0-alpha.1). Current policy is `simple`, so there must be an entry in the changelog with a matching version name for the new release passed to Craft; if not, the command will fail.