feat(performance): make ActivityFramesTracker public
If the Context used in SentryAndroid.init is not an Application
then the ActivityFramesTracker integration is not automatically
added to the available ones.
It might be possible that applications need to init Sentry in
Context#attachBaseContext providing base Context instead of the
current Application.

With this change, integrators can enable ActivityFramesTracker
integration during configuration init.