Add addr_mode attributes to ANR stack frames
The instruction addresses of native stack frames in thread dumps are
relative to the image file from which the code is loaded, and there
are no absolute mapping addresses of images available. So explicitly
inform the Sentry server about the correct images by using a relative
"addr_mode" attribute.

Also add the attribute to the SentryStackFrame class since it was not
yet supported by it. The field documentation is converted from
event.schema.json in the sentry server repo.