Enable symbolication of native stack frames in ANR events (#4061)
* Add native ANR button to sentry-samples-android

Add a button to trigger ANR by holding a lock too long in native
code. This can be used to test native stack frames in ANR events.

* Improve native stack frame parsing

Handle offsets and deleted files, recognize "???" as a marker for
unknown functions. Use named capturing groups for better readability
and editability.

* Add PC value and platform to native stack frames

* Mark JNI method frames as "native"

Use the "native" attribute of stack frames to indicate JNI
invocation frames, like SentryStackTraceFactory does.

* Add debug images to ANR events

The images are parsed from the build ids and filenames in the thread
dump's stack frames.

* Add addr_mode attributes to ANR stack frames

The instruction addresses of native stack frames in thread dumps are
relative to the image file from which the code is loaded, and there
are no absolute mapping addresses of images available. So explicitly
inform the Sentry server about the correct images by using a relative
"addr_mode" attribute.

Also add the attribute to the SentryStackFrame class since it was not
yet supported by it. The field documentation is converted from
event.schema.json in the sentry server repo.

* Add ChangeLog entry for ANR native symbolication

* Fix code formatting, make API 21 level compatible, minor improvements

* Add ADDR_MODE entries to sentry.api

* Update sentry-android-core/src/main/java/io/sentry/android/core/internal/threaddump/ThreadDumpParser.java

* Update sentry-android-core/src/main/java/io/sentry/android/core/internal/threaddump/ThreadDumpParser.java

---------

Co-authored-by: Lauri Alanko <ext-lauri.alanko@supercell.com>
Co-authored-by: Markus Hintersteiner <markus.hintersteiner@sentry.io>
Co-authored-by: Roman Zavarnitsyn <rom4ek93@gmail.com>