Aggregate javadoc using build-logic
This creates a new convention plugin `sentry.javadoc` that
publishes the javadoc to a consumable configuration. The consumable
configuration is then consumed in the root project in order to aggregate
the javadocs of all subprojects.

The producer/consumer configuration is based on this blog post: https://www.liutikas.net/2024/12/11/Together-In-Isolation.html
It should be project isolation compatible.