More accurate, faster spotless config
This PR ensures that spotless is actually running over all our files. It
wasn't before and you can see this generated a massive diff.
There are still a number of exceptions that were suppressed in order to
save time going through the hundreds of them.

In other news, this moves the spotless logic to its own convention
plugin and runs over each source set of each project separately. This
means that most of the tasks should be cached independently.